Restructure feature tests by driver
