+++ tests/Feature/Console/Database/CheckoutProviderCommandTest.php
@@ -1,11 +1,12 @@
 <?php
 
-namespace Payavel\Checkout\Tests\Feature\Console\Commands;
+namespace Payavel\Checkout\Tests\Feature\Console\Database;
 
+use Payavel\Checkout\Tests\Feature\Console\TestCheckoutProviderCommand;
 use Payavel\Orchestration\Tests\Traits\CreatesDatabaseServiceables;
 use Payavel\Orchestration\Tests\Traits\SetsDatabaseDriver;
 
-class DatabaseCheckoutProviderCommandTest extends TestCheckoutProviderCommand
+class CheckoutProviderCommandTest extends TestCheckoutProviderCommand
 {
     use CreatesDatabaseServiceables;
     use SetsDatabaseDriver;
